+++ src/lib.c
@@ -29,4 +29,5 @@
 #include "lib/vec-chr.c"
 #include "lib/vec-lgl.c"
 #include "lib/vec-list.c"
+#include "lib/vendor.c"
 #include "lib/weakref.c"