+++ src/lib/rlang.c
@@ -39,6 +39,7 @@ void r_init_library_stack();
 void r_init_library_sym();
 void r_init_library_vec();
 void r_init_library_vec_chr();
+void r_init_library_vendor();
 
 sexp* r_shared_true;
 sexp* r_shared_false;
@@ -62,6 +63,7 @@ SEXP r_init_library() {
   r_init_library_stack();
   r_init_library_vec();
   r_init_library_vec_chr();
+  r_init_library_vendor();
 
   r_shared_true = r_new_vector(r_type_logical, 1);
   r_mark_precious(r_shared_true);