+++ src/lib/rlang.h
@@ -99,6 +99,7 @@ extern sexp* r_shared_false;
 #include "vec-chr.h"
 #include "vec-lgl.h"
 #include "vec-list.h"
+#include "vendor.h"
 
 sexp* r_init_library();